+++ src/app/guest/friends/page.tsx
@@ -1,9 +1,7 @@
 'use client';
-import { selectedUserAtom } from '@/atoms/atoms';
 import NickNameList from '@/components/guest/NickNameList';
-import { useGetSearchNickName } from '@/services/getSearchNickName';
+import { useGetSearchNickName } from '@/services/guest/getSearchNickName';
 import { NickNameListType } from '@/types/guestBookType';
-import { useAtom } from 'jotai';
 import { useCallback, useState } from 'react';
 import { MdSearch } from 'react-icons/md';
