+++ src/app/guest/page.tsx
@@ -3,8 +3,8 @@ import { userAtom } from '@/atoms/atoms';
 import DeleteAlert from '@/components/guest/DeleteAlert';
 import GuestListItem from '@/components/guest/GuestListItem';
 import useMoveScrollBottom from '@/hooks/useMoveScrollBottom';
-import { useGetGuestBook } from '@/services/getGuestBook';
-import { usePostGuestBook } from '@/services/postGuestBook';
+import { useGetGuestBook } from '@/services/guest/getGuestBook';
+import { usePostGuestBook } from '@/services/guest/postGuestBook';
 import { useAtomValue } from 'jotai';
 import { useRouter } from 'next/navigation';
 import { useCallback, useEffect, useRef, useState } from 'react';
@@ -15,7 +15,7 @@ export default function Guest() {
   const [userInput, setUserInput] = useState('');
   const [modalOpen, setModalOpen] = useState(false);
   const user = useAtomValue(userAtom);
-  const { data: guestBook, isLoading: isGuestBookLoading, error: isGuestBookError } = useGetGuestBook(user);
+  const { data: guestBook } = useGetGuestBook(user);
   const guestBookList = guestBook ?? [];
   const { mutateAsync: postGuestBook } = usePostGuestBook();
   const scrollRef = useMoveScrollBottom(guestBookList);
