+++ src/components/guest/DeleteAlert.tsx
@@ -1,6 +1,6 @@
 import { DeleteAlertProps } from '@/types/guestBookType';
 import ModalWrapper from '../ModalWrapper';
-import { useDeleteGuestBook } from '@/services/deleteGuestBook';
+import { useDeleteGuestBook } from '@/services/guest/deleteGuestBook';
 
 export default function DeleteAlert({ onClose, itemId }: DeleteAlertProps) {
   const { mutateAsync: deleteGuestBook } = useDeleteGuestBook();
