+++ src/services/guest/deleteGuestBook.ts
@@ -1,5 +1,5 @@
 import { DeleteId } from '@/types/guestBookType';
-import { axios } from './instance';
+import { axios } from '../instance';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 const fetchDeleteGuestBook = async (comment_id: DeleteId) => {
