+++ src/services/guest/getGuestBook.ts
@@ -1,7 +1,7 @@
 import { GuestBookListType } from '@/types/guestBookType';
-import { axios } from './instance';
 import { useQuery } from '@tanstack/react-query';
 import { User } from '@/atoms/atoms';
+import { axios } from '../instance';
 
 const fetchGuestBook = async () => {
   const response = await axios.get<GuestBookListType[]>(`guestbook/`);
